requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
cachetools
flake8
python-dotenv>=0.5.1
isort
typing
Pillow
hydra-core
wandb
pytest
opencv-python
imgaug
tqdm
scikit-learn
quantus
pandas
torch
tabulate
torchvision
captum
mpl-scatter-density
scikit-image
h5py
zennit
numpy<=1.20.3
